linux_udev: only return initialized devices during enumeration
The udev enumerator can return devices for which udev has not finished
processing its rules. As udev has not finished handling its rules, the
application may not yet be permitted to open the device.

For these devices a hotplug event will happen later on when udev sends
the "add" event.

Also add a umockdev test for this specific corner case. Note that other
tests need to be updated to work around poor umockdev emulation of the
udev behaviour. umockdev never writes udev database entries to its
/run/udev/data, which means that devices are only considered initialized
when an event has been received.

Signed-off-by: Benjamin Berg <benjamin@sipsolutions.net>